requirements.txt
deprecated~=1.2.14
loguru~=0.7.2
setuptools~=72.1.0
twine
pillow~=10.4.0
pytest~=8.3.3
moviepy~=1.0.3
pydub~=0.25.1
python-dotenv~=1.0.1
freezegun~=1.5.1
Authlib~=1.3.2
pikepdf~=9.3.0
EbookLib~=0.18
mozjpeg-lossless-optimization~=1.1.3
pillow-avif-plugin~=1.4.6
pytest-mock~=3.14.0
uuid_utils~=0.9.0
python-ulid